requirements.txt
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==5.4.0
imutils==0.5.3
intervaltree==3.1.0
ipykernel==6.7.0
ipython==8.0.1
ipython_genutils==0.2.0
Keras-Preprocessing==1.1.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mpmath==1.2.1
msrest==0.6.21
multidict==5.1.0
multiprocessing-logging==0.3.1
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.6.2
nibabel==3.2.1
notebook==6.4.7
numba==0.50.1
numpy==1.21.4
numpydoc==0.7.0
opencv-python==4.5.5.64
pandas==1.3.4
pickleshare==0.7.5
Pillow==8.3.0
PyYAML==5.3.1
scikit-learn==0.23.1
scipy==1.7.2
simplejson==3.16.0
SoundFile==0.10.3.post1
tensorboard==2.4.1
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.1
tensorflow-estimator==2.3.0
tqdm==4.48.0
